move trimmer_utlis tests to separte file
